Update Makefile to use new devcontainer
In the previous commit, we added a VS Code dev container. Before doing
so, we had a home-grown solution akin to a dev container (but not
integrated with the IDE) that used a Makefile. We can delete our old
Dockerfile, and update our Makefile to point to the Dockerfile in our
`.devcontainer` folder. This means:

 * `make test` will still work, and we don't need to change our CI.
 * Anyone who doesn't use VS Code can still run `make test`, `make
   shell`, etc.