Hacky fix for PHP syntax highlighting
Prism.js has a bug that is preventing it from correctly highlighting our
sample PHP code. The bug happens when the code includes a `<?php` tag
and heredoc (`<<<END`) with a single quote (`'`) in it.

The website looks way better with working syntax highlighting, so let's
work around this bug by removing the `'` from the heredoc in our PHP
sample.

The bug has been reported to prism.js:
https://github.com/PrismJS/prism/issues/3571