Add EditorConfig
It would be great to use a code formatter/linter to ensure consistency
within this repo and across all of our seed scripts. But it's
challenging because we want to format code consistently in 15+
languages. [EditorConfig](https://editorconfig.org/) seems like a good
place to start since it's widely used and language-agnostic. We can use
`eclint` in GitHub Actions to enforce code style.

* Use consistent line endings everywhere.
* Use 2-space indent on most files. Enforce this by file type (for
  non-seed code) and for anything in the `seeds/` dir (since we don't
  know file types that might be added in the future).
* Ensure we don't reformat packaged libraries (bootstrap/prismjs).